refactor: tweak bill book add api

+++ src/model/request/bill/bill_book_request.rs
@@ -6,5 +6,5 @@ use rocket_okapi::okapi::schemars;
 #[derive(Debug, PartialEq, Eq, Deserialize, Serialize, JsonSchema)]
 #[allow(non_snake_case)]
 pub struct BillBookRequest {
-    pub billBookTemplateId: i32
+    pub billBookTemplateId: i64
 }
